Update day9.py
Fix for incorrect values on longer ropes.  Rope.touching() now accounts for overlapping nodes, and means we now no longer need the manhattan distance code.  modifying the while loop to an if then wrapping the whole check/move segment in an outer for loop ensures that the nodes follow one another properly.